renderização das comics em userPage (concluído)
