+++ services/UserService.js
@@ -1,6 +1,9 @@
 const db = require("../database/models")
 const jwt = require("jsonwebtoken")
+const database = require('../database/models/index');
+
 const jwtSecret = process.env.JWT_SECRET
+const { Op } = require('sequelize')
 
 const UserService = {
 
@@ -25,6 +28,30 @@ const UserService = {
             { expiresIn: '48h' }    
     )
     return token
+    },
+
+    clientePreferencia: async (id) => {
+        const cliente = await database.Cliente.findOne({
+            where: {
+                id
+            },
+            include: {
+                model: database.Preferencia,
+                required: true
+            }
+        });
+        
+        const { Preferencia } = cliente
+        const preferenciasIds = Preferencia.map( elem => elem.id)
+
+        const clientePreferencias = await database.Preferencia.findAll({
+            where: { 
+                id: { [Op.in]: preferenciasIds}
+            },
+            include: [{ model: database.Comic }]
+        });
+
+        return clientePreferencias
     }
 }
